chore(server):split assistant service
